lint and create marketing-cloud folder
