+++ packages/template-retail-react-app/app/ssr.js
@@ -23,9 +23,9 @@ import {getConfig} from '@salesforce/pwa-kit-runtime/utils/ssr-config'
 import helmet from 'helmet'
 
 import express from 'express'
-import {emailLink} from './marketing-cloud-email-link'
+import {emailLink} from '@salesforce/retail-react-app/app/utils/marketing-cloud/marketing-cloud-email-link'
 
-const ENABLE_SSR_POST = (process.env.ENABLE_SSR_POST || "").toLowerCase() === "true"
+const ENABLE_SSR_POST = (process.env.ENABLE_SSR_POST || '').toLowerCase() === 'true'
 
 const options = {
     // The build directory (an absolute path)
